Various optimizations and bug fixes.

+++ src/main/kotlin/com/criptext/mail/scenes/settings/cloudbackup/data/CloudBackupData.kt
@@ -3,4 +3,4 @@ package com.criptext.mail.scenes.settings.cloudbackup.data
 import java.util.*
 
 data class CloudBackupData(val hasCloudBackup: Boolean, val autoBackupFrequency: Int,
-                           val useWifiOnly: Boolean, val lastModified: Date?, val fileSize: Int)
+                           val useWifiOnly: Boolean, val lastModified: Date?, val fileSize: Long)
